# app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eth-tools API",
    "description": "Self-maintaining runtime for ERC-8004 trustless agents.",
    "license": {
      "name": "MIT"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://eth-tools.dev",
      "description": "Production"
    },
    {
      "url": "https://preview.eth-tools.dev",
      "description": "Preview"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "agents_list",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Max rows to return; clamped to [1, 200].",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 50,
              "maximum": 200,
              "minimum": 1
            }
          },
          {
            "name": "chain",
            "in": "query",
            "description": "Filter by chain name (e.g. 'base') or chain_id (e.g. '8453').",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque keyset cursor from a prior response's next_cursor.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List indexed agents (keyset-paginated)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentList"
                }
              }
            }
          },
          "400": {
            "description": "invalid cursor or query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/search": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "agents_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Filtered agent list (non-paginated, hard cap 200)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentList"
                }
              }
            }
          },
          "400": {
            "description": "invalid filter (bad chain / bad owner hex)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{chain}/{agent_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "agents_get_one",
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Chain name (e.g. 'base') or chain_id (e.g. '8453')",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "description": "uint256 as decimal string",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a single agent by chain and id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "400": {
            "description": "invalid agent id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "agent or chain not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentDetail": {
        "type": "object",
        "description": "Concrete instantiation of `OneEnvelope<AgentDto>` — see `AgentList`.",
        "required": [
          "data",
          "staleness_ms",
          "source"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AgentDto"
          },
          "source": {
            "type": "string",
            "example": "db"
          },
          "staleness_ms": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AgentDto": {
        "type": "object",
        "required": [
          "chain",
          "chain_id",
          "agent_id",
          "owner",
          "registered_at",
          "updated_at"
        ],
        "properties": {
          "agent_id": {
            "type": "string",
            "description": "uint256 as decimal string",
            "example": "42"
          },
          "agent_uri": {
            "type": [
              "string",
              "null"
            ]
          },
          "agent_wallet": {
            "type": [
              "string",
              "null"
            ],
            "description": "0x-prefixed; cleared on ownership Transfer (spec gotcha)"
          },
          "chain": {
            "type": "string",
            "example": "base"
          },
          "chain_id": {
            "type": "integer",
            "format": "int64",
            "example": 8453
          },
          "owner": {
            "type": "string",
            "description": "0x-prefixed 20-byte hex"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AgentList": {
        "type": "object",
        "description": "Concrete instantiation of `ListEnvelope<AgentDto>` so the OpenAPI spec\nemits a stable `AgentList` schema name (utoipa generates synthetic names\nlike `ListEnvelope_AgentDto` for raw generic instantiations).",
        "required": [
          "data",
          "staleness_ms",
          "source"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentDto"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque; pass verbatim as ?cursor= to page"
          },
          "source": {
            "type": "string",
            "example": "db"
          },
          "staleness_ms": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "Top-level error envelope returned by every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayload"
          }
        }
      },
      "ApiErrorPayload": {
        "type": "object",
        "required": [
          "code",
          "policy_version",
          "evaluator"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "AGENT_NOT_FOUND"
          },
          "evaluator": {
            "type": "string",
            "example": "api.lookup"
          },
          "override_hint": {
            "type": [
              "string",
              "null"
            ]
          },
          "policy_version": {
            "type": "string",
            "example": "v1"
          }
        }
      },
...
      "SearchFilters": {
        "type": "object",
        "properties": {
          "chain": {
            "type": [
              "string",
              "null"
            ],
            "description": "Chain name (`base`) or chain_id as decimal string (`8453`). Omit for any."
          },
          "has_endpoint": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "has_manifest": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "owner": {
            "type": [
              "string",
              "null"
            ],
            "description": "0x-prefixed 20-byte hex owner address."
          }
        }
      },
      "SearchRequest": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Clamped to [1, 200]; default 50."
          },
          "query": {
            "type": [
              "string",
              "null"
            ],
            "description": "Substring matched against `agent_uri` ILIKE. Omit for \"any URI\"."
          }
        }
      },
...
    }
  },
...
}